python/libccv/web_cam_fastapi/requirements.txt
fastapi
uvicorn
opencv-python-headless
jinja2
numpy
requests
pyzmq
